# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: user-service
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /auth/signup:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                provider: { type: string, enum: [local, google, github] }
              required: [email]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  verification_token: { type: string }
  /auth/signin:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
              required: [email, password]
      responses:
        "200":
          description: ok
  /auth/code-verification:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                code: { type: string }
              required: [token, code]
      responses:
        "200":
          description: ok
  /auth/restore:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
              required: [email]
      responses:
        "200":
          description: ok
  /users/me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: ok }
  /users:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: ok }
  /users/{id}:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: ok }
    delete:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: ok }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer